amdgpu: Add missing specialization and move format mapping data to types
